+++ service/user_register_service.go
@@ -23,7 +23,7 @@ func (service *UserRegisterService) Valid() *serializer.Response {
 	}
 
 	// 不允许昵称相同
-	count := 0
+	count := int64(0)
 	model.DB.Model(&model.User{}).Where("nickname = ?", service.Nickname).Count(&count)
 	if count > 0 {
 		return &serializer.Response{
